Silicon/ARM/NeoverseN1Soc: NOR flash library for N1Sdp
Add NOR flash library, this library provides APIs for getting the list
of NOR flash devices on the platform.

This flash is shared between AP core and System Control Processor. The
lower addresses are used to store SCP and AP boot images and higher
addresses will be used for variable storage.

Signed-off-by: sahil <sahil@arm.com>
Change-Id: Idf0cb6990cde233b6c6cef8a36aa342c44257325